+++ Network/INetworkManager.cs
@@ -6,6 +6,7 @@ internal interface INetworkManager : IDisposable
     bool IsHost { get; }
     bool IsSingleplayer { get; }
     bool IsServer { get; }
+    void RegisterLoggerSettings();
     void Update();
     void BroadcastStation(string stationGuid, bool value);
 }